backend/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
corsheaders==3.14.0